backend/requirements.txt
# --- Core Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1

# --- Database & ORM ---
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Security & Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pinned: passlib 1.7.4 incompatible with bcrypt>=4.1
python-multipart==0.0.9

# --- Data Validation ---
pydantic[email]==2.6.3
pydantic-settings==2.2.1

# --- Utilities ---
python-dotenv==1.0.1